removed server package and moved server.go file to the router package
