+++ src/router/router.go
@@ -4,7 +4,6 @@ import (
 	"os"
 
 	"github.com/kevinanielsen/go-fast-cdn/src/middleware"
-	"github.com/kevinanielsen/go-fast-cdn/src/server"
 	"github.com/kevinanielsen/go-fast-cdn/ui"
 )
 
@@ -13,13 +12,13 @@ import (
 func Router() {
 	port := ":" + os.Getenv("PORT")
 
-	s := server.NewServer(
-		server.WithPort(port),
-		server.WithMiddleware(middleware.CORSMiddleware()),
+	s := NewServer(
+		WithPort(port),
+		WithMiddleware(middleware.CORSMiddleware()),
 	)
 
 	// Add all the API routes
-	AddApiRoutes(s.Engine)
+	s.AddApiRoutes()
 
 	// Add the embedded ui routes
 	ui.AddRoutes(s.Engine)
